Added pre-condition checks to business services

+++ app/services/complete_task.rb
@@ -5,7 +5,7 @@ class CompleteTask
   
   def run
     Task.find(@task_id).tap do |task|
-      # May only happen if not yet completed
+      raise ArgumentError.new("Task is already completed") if task.completed
       TaskPositionService.remove(task)
       task.completed = true
       task.save!
